pyproject.toml
[project]
name = "cloud_storage"
authors = [
    { name = "AgriTheory", email = "support@agritheory.com"}
]
description = "Frappe App for integrating with cloud storage applications"
requires-python = ">=3.10"
readme = "README.md"
dynamic = ["version"]
dependencies = [
    "boto3~=1.26.38",
    "botocore~=1.29.38",
    "python-magic~=0.4.27",
]

[build-system]
requires = ["flit_core >=3.4,<4"]
build-backend = "flit_core.buildapi"

[tool.black]
line-length = 99

[tool.isort]
line_length = 99
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
indent = "\t"
